+++ src/main/java/it/unibz/infosec/examproject/product/domain/ManageProducts.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.sql.Date;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -20,7 +21,7 @@ public class ManageProducts {
         this.manageUsers = manageUsers;
     }
 
-    private Product validateProduct (Long id) {
+    private Product validateProduct(Long id) {
         final Optional<Product> maybeProduct = productRepository.findById(id);
         if (maybeProduct.isEmpty())
             throw new IllegalArgumentException("Product with id '" + id + "' does not exist yet!");
@@ -35,23 +36,26 @@ public class ManageProducts {
         return validateProduct(id);
     }
 
-    public Product updateProduct (Long id, String name, int cost) {
-
-        Product product = validateProduct(id);
-
+    public Product updateProduct(Long id, Long vendorId, String name, int cost) {
+        final Product product = validateProduct(id);
+        if (!vendorId.equals(product.getVendorId())) {
+            throw new IllegalArgumentException("Only the owner of a product can update it");
+        }
         product.setName(name);
         product.setCost(cost);
-
         return productRepository.save(product);
     }
 
-    public Product deleteProduct (Long id) {
-
-        Product product = validateProduct(id);
-
+    public Product deleteProduct(Long id, Long vendorId) {
+        final Product product = validateProduct(id);
+        if (!vendorId.equals(product.getVendorId())) {
+            throw new IllegalArgumentException("Only the owner of a product can delete it");
+        }
         productRepository.delete(product);
-
         return product;
     }
 
+    public List<Product> getByVendor(Long vendorId) {
+        return productRepository.findByVendorId(vendorId);
+    }
 }
