+++ src/main/java/it/unibz/infosec/examproject/review/domain/ManageReviews.java
@@ -1,10 +1,12 @@
 package it.unibz.infosec.examproject.review.domain;
 
 import it.unibz.infosec.examproject.product.domain.ManageProducts;
+import it.unibz.infosec.examproject.user.domain.ManageUsers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.sql.Date;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -13,11 +15,13 @@ public class ManageReviews {
     private final ReviewRepository reviewRepository;
     private SearchReviews searchReviews;
     private final ManageProducts manageProducts;
+    private final ManageUsers manageUsers;
 
     @Autowired
-    public ManageReviews(ReviewRepository reviewRepository, ManageProducts manageProducts) {
+    public ManageReviews(ReviewRepository reviewRepository, ManageProducts manageProducts, ManageUsers manageUsers) {
         this.reviewRepository = reviewRepository;
         this.manageProducts = manageProducts;
+        this.manageUsers = manageUsers;
     }
 
     private Review validateReview(Long id) {
@@ -28,11 +32,8 @@ public class ManageReviews {
         return maybeReview.get();
     }
 
-    //TODO: validate author Id
     public Review createReview(String title, String description, int stars, Date datePublishing, Long productId, Long replyFromReviewId, Long author) {
-
-        Long replyReviewId = replyFromReviewId == 0 ? 0 : validateReview(replyFromReviewId).getId();
-
+        final Long replyReviewId = replyFromReviewId == null || replyFromReviewId == 0 ? 0 : validateReview(replyFromReviewId).getId();
         return reviewRepository.save(new Review(
                 title,
                 description,
@@ -40,25 +41,42 @@ public class ManageReviews {
                 datePublishing,
                 manageProducts.readProduct(productId).getId(),
                 replyReviewId,
-                author)
-        );
+                manageUsers.readUser(author).getId()
+        ));
     }
 
     public Review readReview(Long id) {
         return validateReview(id);
     }
 
-    public Review updateReview (Long id, String title, String description) {
+    public Review updateReview(Long id, Long author, String title, String description) {
         final Review review = validateReview(id);
+        if (!review.getAuthor().equals(author)) {
+            throw new IllegalArgumentException("Only the author of a review can update it");
+        }
         review.setTitle(title);
         review.setDescription(description);
         return reviewRepository.save(review);
     }
 
-    public Review deleteReview (Long id) {
+    public Review deleteReview(Long id, Long author) {
         final Review review = validateReview(id);
+        if (!review.getAuthor().equals(author)) {
+            throw new IllegalArgumentException("Only the author of a review can delete it");
+        }
         reviewRepository.delete(review);
         return review;
     }
 
+    public List<Review> getByCustomer(Long customerId) {
+        return reviewRepository.findByAuthor(customerId);
+    }
+
+    public List<Review> getByProduct(Long productId) {
+        return reviewRepository.findByProductId(productId);
+    }
+
+    public List<Review> getReplies(Long reviewId) {
+        return reviewRepository.findByReplyFromReviewId(reviewId);
+    }
 }
